extract mapper
