+++ src/modules/aws_appsync/mappers/index.ts
@@ -1,0 +1,1 @@
+export * from './graphql_api';
